api/requirements.txt
-r ../shared.txt
fastapi
uvicorn
python-multipart
torch
torchvision
transformers
Pillow
joblib
scipy
git+https://github.com/openai/CLIP.git